AI_Agents/Candidate_AI_Advocate/requirements.txt
langchain 
openai 
pinecone-client 
python-dotenv 
streamlit 
google-api-python-client
